requirements.txt
pandas>=1.3.0
geopandas>=1.0.1
numpy>=1.21.0
requests>=2.25.0
pyogrio>=0.4.0
shapely>=2.0.0
openpyxl>=3.0.10
python-dotenv==1.0.0
typing_extensions
matplotlib>=3.0.0
